izmene nakon dodate authentifikacije

+++ app/Http/Controllers/PostsController.php
@@ -6,6 +6,7 @@ use App\Models\Category;
 use App\Models\Post;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostsController extends Controller
 {
@@ -16,12 +17,21 @@ class PostsController extends Controller
      */
     public function index() //get ruta - jer prihvatamo podatke iz baze
     {
-        //vraca sve postove iz baze
+        // vraca sve postove iz baze
         $posts = Post::all();
         // return response()->json($posts);
         return view('blog.index',[
             'posts'=>$posts
         ]);
+
+        //izmena ako vraca samo postove korisnika koji ih je postavio
+        // $user = Auth::user();
+        // $user_posts = $user->posts;
+        // return view('blog.index',[
+        //     'posts'=>$user_posts
+        // ]);
+        //ili ako hocemo da prikazujemo sve na home a preko /my samo za ulogovanog
+        // mozemo da izmenimo show 
     }
 
     /**
@@ -67,12 +77,28 @@ class PostsController extends Controller
      */
     public function show($postid) //get ruta - jer prihvatamo podatke iz baze
     {
-        $post = Post::find($postid);
+        // $post = Post::find($postid);
 
-        return view('blog.show', [
-            'post' => $post,
+        // return view('blog.show', [
+        //     'post' => $post,
+        // ]); 
+        if($postid=='my'){
+            //izmena ako vraca samo postove korisnika koji ih je postavio
+            $user = Auth::user();
+            $user_posts = $user->posts;
+            return view('blog.index',[
+                'posts'=>$user_posts
+            ]);
+        }else{
 
-        ]); 
+            //izmena ako za autentifikaciju
+            $post = Post::find($postid);
+            $user = Auth::id();
+            return view('blog.show', [
+                'post' => $post,
+                'auth_user'=>$user
+            ]); 
+        }
     }
 
     /**
@@ -83,14 +109,28 @@ class PostsController extends Controller
      */
     public function edit($postid) //get ruta - jer zelimo da prikazemo edit form view
     {
+        // $post = Post::find($postid);
+        // $categories = Category::all();
+
+        // return view('blog.edit', [
+        //     'post' => $post,
+        //     'categories'=>$categories,
+        //     ]);
+        
+        //izmena sa autentifikacijom
+        $user = Auth::user();
         $post = Post::find($postid);
-        $categories = Category::all();
+        if($user->id == $post->user->id){
+            $categories = Category::all();
 
         return view('blog.edit', [
             'post' => $post,
             'categories'=>$categories,
 
-            ]);
+        ]);
+        }else{
+            return redirect('blog/');
+        }
     }
 
     /**
@@ -121,10 +161,18 @@ class PostsController extends Controller
      */
     public function destroy($postid) //delete ruta - jer zelimo da obrisemo iz baze jedan red
     {
+        // $post = Post::find($postid);
+        // $post->delete();
+        // return redirect('/blog');
+
+        //izmena za autentifikaciju
         $post = Post::find($postid);
-
-        $post->delete();
-
+        if($post->user->id == Auth::id()){
+            $post->delete();
+        }
         return redirect('/blog');
+        
+        
+
     }
 }
